Improve variable names and documentation in `core.py`
Renamed `w` to `ws_global` to clearly indicate its role as a collection of hyperparameters. In addition to that, improved variable names like `border` to `triangle_edges` and `boundary` to `triangle_edge` to be more accurate and descriptive. Some explanatory comments were also added to further increase the readability and comprehensibility of the code.