Refactor class name and add documentation
Class "Hierarchy_position_data_model" has been renamed to "Hierarchical_position_data_model" for consistency of symbol names. Additionally, documentation was added to explain its role in encoding the topological position in a subdivision hierarchy. The instances of this class in the code were also updated to reflect the new name.